Make Emacs term-named function to create multiple Term mode instances
This function also gets rid of a lot of warnings while byte-compiling .emacs.

As for the function, there is a new interactive function that solves the problem
of not having multiple term-mode windows by allowing the invoker to explicitly
give a term-mode buffer a new name (causing it to not simply jump to an existing
buffer).