Rearchitect Emacs font highlighting
This new approach no longer relies on manual function calls that differ from
light to dark background, and from version to version.  Instead, we build a list
of all possible face configurations, and conditionally apply them in a generic
manner.  This is to make it easier to expand font coloration to include 256
color terminals.