Fully enable customized syntax coloring for 256 color terminals.
This new code supports XTerm as long as the mode is something that indicates 256
or more colors.  It also supports Konsole, which required a bit of hacking
around with environment variables, so it's a bit more delicate.  MacOS and
Cygwin's MinTTY are unknowns right now.  Backward compatibility may also have
broken.  Deal with that later...