Fix Emacs desktop loading on the terminal
THIS COULD NOT BE STUPIDER.  In order for `desktop-read' to restore the frameset
that was saved in the desktop, it must call `desktop-restoring-frameset-p',
which in turn checks the function `display-graphic-p'.  This essentially means
that, even though restoring the frameset works on the TTY (providing
`desktop-restore-forces-onscreen' is NIL), it NEVER runs.  So we get the full
desktop, but none of the frames.  FIXING THIS NOW.

Secondly, we were calling `desktop-change-dir' which left the default desktop
path in tact (pointing to the wrong place), and forced a pointless desktop
shutdown before loading the desktop we want.