Comprehensively re-work Emacs configuration for multiple clients
After using emacsclient from multiple machines with different terminal types,
some issues in the configuration became apparent.  First and foremost, the
colors that are set should not apply to all frames.  This change puts hooks into
the frame functions to only set the color scheme when a frame is created, using
that frames settings to determine if a light or dark color scheme should be
used.

!!! There is a known issue where frames loaded from a desktop don't come up with
    the right color. !!!

This change also fixes up the clang-format usage, and clears up some
documentation lint.  There are probably more issues, but I've tested this with
Emacs 20 and 23, and things work OK.