Be more consistent with Emacs function variable treatment
In particular, it is more efficient to define anonymous functions with the
"function" special form, and #' is shorthand for this, so that's what we're
using.  This #' syntax is also now applied to named function symbols being
passed around as well.  There should be no deviation in the code any more.