Shuffle GDB initialization files around.
This moves the Palm GDB initialization file out of the way, and create an empty
one for now.  There is all this Holly-Mail stuff in there that's not relevant at
all to me, and probably a bunch of the settings are super old.  It should be
possible to conditionally load these dev-specific environments from an agnostic
GDB init script.