Fix a long-standing annoyance in Emacs GDB behavior
The overlay for `gdb-mode' (`gud-mode' as a whole, really) would remain in place
after the GDB process exited sometimes, which was pretty annoying.  This change
advises the sentinel function for the GDB process to remove the overlay (that I
added) at the same time it removes the arrow next to the source line
(default-behavior).