Fix Emacs syntax highlighting under the highlight region.
Before this change, highlighting text would remove any bold/italics/underline
from the highlighted text.  This is because `modify-face' does not provide any
means of saying "don't care" with a face setting.  We should be using
`set-face-attributes' in code, but that doesn't exist in older Emacs version, so
we have to write our own compatible version.