Enable potential optimization of Emacs initialization file version checks.
This change introduces a setting, `lock-dotfile-version', which allows the
compilation of emacs.el to assume a fixed major version.  This enables the
compilation to remove all major version checks, and simply embed the correct
code.  If this is NIL, compiled files will support any Emacs version from 19
onward.

THIS IS LIKELY A BREAKING CHANGE: re-test backward compatibility.