Fix some Emacs font-lock errors
Problem 1: setting a nil background or foreground left the background or
foreground we were trying to replace.  If a pre-defined face has a background
color you don't like, and you want to just make it have no background, you
couldn't do that before today.

Problem 2: Version 24+ fonts weren't being used.  This was because an `if' was
used instead of a `where'.