Allow emacs.el to run "stand-alone" without modification
Because of recent changes, it was no longer possible to simply copy the emacs.el
file to another location and use it directly.  This was because the file
required proximity to the custom packages in elisp/ to be in the same directory
during compilation.  Now, it the file senses that it is the only .el file in its
directory, it avoids loading those extra files, removing the errors.