Provide Doxygen syntax coloration in Emacs for C/C++
This addresses a long-standing annoyance of mine with Emacs's documentation
highlighting.  Why does JavaDoc get highlights and C/C++ does not?  `cc-mode'
actually makes this pretty easy to do, just by adding a constant or two, and
changing a style parameter.

This also fixes a bug in our `eval-after-load' logic, where we added hooks in
the statements that were evaluated.  The hooks sometimes are added after they
would have run, so for the first buffer (i.e., the one that potentially causes
the `eval-after-load'), the hooks don't run at all.  In all cases, this was
affecting `cc-mode', which comes with a handy `c-initialization-hook' that we
can run instead.