Additional fixes to patch up Emacs desktop loading in Emacs daemon
There were two problems.  First, desktop-restore would kill the dummy frame that
keeps the daemon alive.  This caused C-x 5 0 on the last frame to terminate the
daemon, which is not what we want.  Second, UI frames were not showing up for me
over SSH because it was complaining about not finding the dummy DISPLAY setting
that SSH makes (e.g., "localhost:12.0"), which it really has no business using.
I've dropped this setting from the function that saves and loads frameset data.