Fix up ediff issues after all the recent changes.
For starters, the ediff control frame had a menu bar on it, which ediff
explicitly turns off, and we subsequently re-enabled.  Secondly, the dark
background diff coloration was still not great, so I patched that up.  Lastly,
for some reason, the mouse gets thrown off the screen, so I disabled the setting
that tries to place the cursor in the control frame.