Fix a but with xdg-open after trying to pull more KDE settings into FVWM
Basically, when xdg-open finds KDE_FULL_SESSION, it assumes we're definitely
using KDE, and then it looks for other variables.  It doesn't find them, and
falls back to a method that doesn't seem to work on my workstation.  The fix is
to explicitly set XDG_CURRENT_DESKTOP to 'X-Generic' for FVWM (and other
"generic" friends).